backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0
python-cors==1.7.0
ollama-python==0.1.7
transformers==4.35.2
torch==2